+++ framework/domain/src/test/java/is/codion/framework/domain/entity/DefaultEntityDefinitionTest.java
@@ -140,7 +140,7 @@ public class DefaultEntityDefinitionTest {
 		assertTrue(defaultSelectAttributes.contains(Detail.MASTER_CODE_NON_DENORM));
 		assertFalse(defaultSelectAttributes.contains(Detail.MASTER_CODE));
 		assertTrue(defaultSelectAttributes.contains(Detail.MASTER_VIA_CODE_FK));
-		assertTrue(defaultSelectAttributes.contains(Detail.INT_VALUE_LIST));
+		assertTrue(defaultSelectAttributes.contains(Detail.INT_ITEMS));
 		assertFalse(defaultSelectAttributes.contains(Detail.INT_DERIVED));
 		assertFalse(defaultSelectAttributes.contains(Detail.BYTES));
 	}