+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.21
 ### is.codion.framework.domain
 - AuditColumn removed, unused.
+- AbstractAttributeDefinition, Item based attributes now use a custom comparator, sorting by caption.
 ### is.codion.plugin.imagepanel
 - NavigableImagePanel.isWithinImage() bug fixed, now returns false if image is null.